+++ app/src/androidTest/java/edu/up/weininge/catan/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package edu.up.weininge.testgui;
+package edu.up.weininge.catan;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
