+++ app/src/main/java/edu/up/weininge/catan/HexagonGrid.java
@@ -1,4 +1,4 @@
-package edu.up.weininge.testgui;
+package edu.up.weininge.catan;
 
 import android.content.Context;
