+++ app/src/main/res/layout/activity_main.xml
@@ -373,7 +373,7 @@
         app:layout_constraintStart_toStartOf="@+id/button_scoreboard"
         app:layout_constraintTop_toBottomOf="@+id/button_scoreboard" />
 
-    <edu.up.weininge.testgui.boardSurfaceView
+    <edu.up.weininge.catan.boardSurfaceView
         android:id="@+id/boardSurfaceView"
         android:layout_width="0dp"
         android:layout_height="0dp"
